requirements.txt
browser-use>=0.2.0
google-genai>=0.3.0
python-dotenv>=1.0.0
requests>=2.31.0
supabase>=2.3.0
